Flask/iHome/requirements/install.txt

pymysql
redis
flask-blueprint
flask-sqlalchemy
flask-session
flask-script